requirements.txt
attrs==23.1.0 ; python_version >= "3.9" and python_version < "4.0"
backoff==1.11.1 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.6 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
dicom-validator==0.3.4 ; python_version >= "3.9" and python_version < "4.0"
dotty-dict==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
flywheel-gear-toolkit==0.6.10 ; python_version >= "3.9" and python_version < "4.0"
flywheel-gears==0.2.1 ; python_version >= "3.9" and python_version < "4.0"
flywheel-sdk==16.19.2 ; python_version >= "3.9" and python_version < "4.0"
fw-classification==0.3.5 ; python_version >= "3.9" and python_version < "4.0"
fw-core-client==1.1.2 ; python_version >= "3.9" and python_version < "4.0"
fw-file==1.4.2 ; python_version >= "3.9" and python_version < "4.0"
fw-http-client==1.3.2 ; python_version >= "3.9" and python_version < "4.0"
fw-meta==2.1.1 ; python_version >= "3.9" and python_version < "4.0"
fw-utils==4.3.4 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
jsonschema==3.2.0 ; python_version >= "3.9" and python_version < "4.0"
memoization==0.4.0 ; python_version >= "3.9" and python_version < "4"
natsort==8.4.0 ; python_version >= "3.9" and python_version < "4.0"
nibabel==3.2.2 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.25.1 ; python_version >= "3.9" and python_version < "4.0"
packaging==21.3 ; python_version >= "3.9" and python_version < "4.0"
pathvalidate==2.5.2 ; python_version >= "3.9" and python_version < "4.0"
pydantic==1.10.11 ; python_version >= "3.9" and python_version < "4.0"
pydicom==2.4.1 ; python_version >= "3.9" and python_version < "4.0"
pyparsing==3.1.0 ; python_version >= "3.9" and python_version < "4.0"
pyrsistent==0.19.3 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
pytz-deprecation-shim==0.1.0.post0 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
rfc3987==1.3.8 ; python_version >= "3.9" and python_version < "4.0"
ruamel-yaml-clib==0.2.7 ; platform_python_implementation == "CPython" and python_version < "3.12" and python_version >= "3.9"
ruamel-yaml==0.17.32 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
typer==0.4.2 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.9" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "4.0"
tzlocal==4.3.1 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.0.4 ; python_version >= "3.9" and python_version < "4.0"
